Improved coding structure of UI
