+++ src/components/ChannelSettingsContainer.tsx
@@ -5,6 +5,9 @@ import { MIDIDataTable } from "../js/MIDIDataTable";
 import GraviBoardUI from "./classSpecific/GraviBoardUI";
 import PolyDrumUI from "./classSpecific/PolyDrumUI";
 import PresetUI from "./PresetUI";
+import ColorSettingsUI from "./ColorSettingsUI"
+import KeySettingsUI from "./KeySettingsUI"
+
 import * as utils from "../js/solidjs/ComponentUtils.js";
 import * as ui from "./UIElements.jsx"
 
@@ -18,15 +21,9 @@ export default function SetupContainer( props: {channel: number}) {
     const [selectToy, setSelectToy] = createSignal(false);
     const [toyType, setToyType] = createSignal(0);
     const [toyName, setToyName] = createSignal("EmptyToy");
-    const [numberOfKeys, setNumberOfKeys] = createSignal(12);
-    const [startKey, setStartKey] = createSignal(12);
-    const [collapsNote, setCollapsNote] = createSignal(true);
-    //Colors
-    const [colorSelection, setColorSelection] = createSignal(0);
-    const [colorSelectionName, setColorSelectionName] = createSignal("Fill Color");
-    const [fillColor, setFillColor] = createSignal<RGBA>({ r:0, g:0, b:0, a:0});
-    const [strokeColor, setStrokeColor] = createSignal<RGBA>({ r:0, g:0, b:0, a:0});
-    const [accentColor, setAccentColor] = createSignal<RGBA>({ r:0, g:0, b:0, a:0});
+    // const [numberOfKeys, setNumberOfKeys] = createSignal(12);
+    // const [startKey, setStartKey] = createSignal(12);
+    // const [collapsNote, setCollapsNote] = createSignal(true);
 
     createEffect(() => {
         if(useEffect()) {
@@ -40,13 +37,13 @@ export default function SetupContainer( props: {channel: number}) {
         } 
     })
 
-    const ToyChanged = () => {
-        // Handle the event...
-        // console.log("DEFAULT UI event");
-        setUseEffect(false);
-        UpdateUIValues();
-        setUseEffect(true);
-    };
+    // const ToyChanged = () => {
+    //     // Handle the event...
+    //     // console.log("DEFAULT UI event");
+    //     setUseEffect(false);
+    //     UpdateUIValues();
+    //     setUseEffect(true);
+    // };
 
     //Update UI data
     function UpdateUIValues() {
@@ -54,17 +51,9 @@ export default function SetupContainer( props: {channel: number}) {
         if (typeof window !== 'undefined') {
 
             if(toy != undefined) {
+                setUseEffect(false);
                 setToyName(toy.toyName);
-                setNumberOfKeys(toy.numberOfKeys);
-                setStartKey(toy.startKey);
-                setCollapsNote(toy.useRegExp);
-
-                var mColor: RGBA = toy.GetPaperColor(toy.fillColor);
-                var sColor: RGBA = toy.GetPaperColor(toy.strokeColor);
-                var aColor: RGBA = toy.GetPaperColor(toy.accentColor);
-                setFillColor({r:mColor.r, g:mColor.g , b:mColor.b, a:mColor.a});
-                setStrokeColor({r:sColor.r, g:sColor.g , b:sColor.b, a:sColor.a});
-                setAccentColor({r:aColor.r, g:aColor.g , b:aColor.b, a:aColor.a});
+                setUseEffect(true);
             }
         } else setToyName("Empty");
     }
@@ -73,20 +62,20 @@ export default function SetupContainer( props: {channel: number}) {
         // console.log("UPDATE toy values");
         if (typeof window !== 'undefined') {
             //Remove old children
-            toy.RemoveChildrenFromLayer();
+            // toy.RemoveChildrenFromLayer();
 
-            toy.numberOfKeys = numberOfKeys();
-            toy.startKey = startKey();
-            toy.useRegExp = collapsNote();
+            // toy.numberOfKeys = numberOfKeys();
+            // toy.startKey = startKey();
+            // toy.useRegExp = collapsNote();
 
-            toy.SetPaperColor(toy.fillColor, fillColor().r, fillColor().g, fillColor().b, fillColor().a);
-            toy.SetPaperColor(toy.strokeColor, strokeColor().r, strokeColor().g, strokeColor().b, strokeColor().a);
-            toy.SetPaperColor(toy.accentColor, accentColor().r, accentColor().g, accentColor().b, accentColor().a);
+            // toy.SetPaperColor(toy.fillColor, fillColor().r, fillColor().g, fillColor().b, fillColor().a);
+            // toy.SetPaperColor(toy.strokeColor, strokeColor().r, strokeColor().g, strokeColor().b, strokeColor().a);
+            // toy.SetPaperColor(toy.accentColor, accentColor().r, accentColor().g, accentColor().b, accentColor().a);
 
-            if(toyType() != 0) {
-                toy.SetupMIDIReceiver(collapsNote());
-                toy.SetupKeyboard();
-            }
+            try {
+                // toy.SetupMIDIReceiver(collapsNote());
+                // toy.SetupKeyboard();
+            } catch{}
         }
     }
 
@@ -101,8 +90,16 @@ export default function SetupContainer( props: {channel: number}) {
         UpdateUIValues();
         prevToyType = toyType();
     }
+    function UpdateComponent() {
+        UpdateUIValues();
+    }
+
     function LoadToy() {
-        toy = utils.InitToy(channel, toy, ToyChanged);
+        var t = utils.InitToy(channel, toy, UpdateComponent);
+        if(toy != t) {
+            toy = t;
+            UpdateUIValues();
+        }
     }
 
     //true or false, show Toy Selection or Toy Editing Panel
@@ -118,20 +115,6 @@ export default function SetupContainer( props: {channel: number}) {
         setSelectToy(false);
         NewToy();
     }
-    //What kind of color do you want to edit?
-    function UpdateColorSelection(value: number) {
-        var calc = colorSelection();
-        calc += value;
-        if(calc < 0) calc = 2;
-        if(calc > 2) calc = 0;
-
-        switch(calc) {
-            case 0: setColorSelectionName("Fill Color"); break;
-            case 1: setColorSelectionName("Stroke Color"); break;
-            case 2: setColorSelectionName("Accent Color"); break;
-        }
-        setColorSelection(calc);
-    }
 
     //Takes all UI functions and returns it in one big package
     function RenderUI() {
@@ -146,187 +129,25 @@ export default function SetupContainer( props: {channel: number}) {
             else {
                 return(
                     <div class="noSelect">
-                        {RenderPresetUI()}
+                        {/* {RenderPresetUI()} */}
+                        <PresetUI channel={channel} />
                         <br></br>
                         {RenderSpecificUISettings()}
                         <br></br>
-                        {ui.DetailsFillerCenter("Color Settings", RenderColorSettings())}
+                        <ColorSettingsUI channel={channel} />
                         <br></br>
-                        {ui.DetailsFillerCenter("Key Settings", RenderKeySettings())}
+                        <KeySettingsUI channel={channel} />
+                        {/* <ui.DetailsFillerCenter summeryName={"Key Settings"} content={RenderKeySettings} /> */}
                     </div>
                 )
             }
-        }
-    }
-    //Key note settings
-    function RenderKeySettings() {
-        return(
-            <>
-                <ui.NumberSliderUIElement 
-                    name={"Keys"}
-                    minMaxStep={[1,100,1]}
-                    value={numberOfKeys()}
-                    onChange={setNumberOfKeys}
-                />
-                {/* <div class="flexContainer">
-                    <div>Keys</div>
-                    <div class="flexContainer">
-                        <NumberSliderCombo
-                            minMaxStep={[1, 100, 1]}
-                            value={numberOfKeys()}
-                            onChange={setNumberOfKeys}
-                        />
-                    </div>
-                </div> */}
-                {RenderStartKeySetting()}
-                <ui.CheckboxUIElement 
-                name="Collapse Notes"
-                checked={collapsNote()}
-                onChange={setCollapsNote}
-                />
-            </>
-        )
-    }
-    //Retusn more or less UI, depends on a settings
-    function RenderStartKeySetting() {
-        //If collaps note is true, there is no need to set a start key, hide it
-        if(collapsNote() == true) {
-            return (
-                // Sorry, nothing
-                <div></div>
-            )
-        } else {
-            return (
-                <ui.NumberSliderUIElement 
-                    name={"Start Key (" + MIDIDataTable.MIDINoteToString(startKey()) + ")"}
-                    minMaxStep={[1,100,1]}
-                    value={startKey()}
-                    onChange={setStartKey}
-                />
-            )
-        }
-    }
-    //Renders basic Color UI (Buttons and Name)
-    function RenderColorSettings() {
-        return (
-            <>
-                <div class="flexContainer">
-                    <div class="flexContainer">
-                        <ui.ButtonIcon
-                            icon="material-symbols:chevron-left"
-                            onClick={() => UpdateColorSelection(-1)}
-                            width={30}
-                        />
-                        <ui.ButtonIcon
-                            icon="material-symbols:chevron-right"
-                            onClick={() => UpdateColorSelection(1)}
-                            width={30}
-                        />
-                        <h3 class="marginLeft20">{colorSelectionName()}</h3>
-                    </div>
-                </div>
-                <br></br>
-                {RenderColorSettingsSelection()}
-            </>
-        )
-    }
-    //Returns fillColor, strokeColor or accentColor UI
-    function RenderColorSettingsSelection() {
-        switch(colorSelection()) {
-            //fillColor
-            case 0: return (
-            <div>
-                <ui.NumberSliderUIElement 
-                    name={"Red"}
-                    minMaxStep={[0,255,1]}
-                    value={fillColor().r}
-                    onChange={(value) => setFillColor({ ...fillColor(), r: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Green"}
-                    minMaxStep={[0,255,1]}
-                    value={fillColor().g}
-                    onChange={(value) => setFillColor({ ...fillColor(), g: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Blue"}
-                    minMaxStep={[0,255,1]}
-                    value={fillColor().b}
-                    onChange={(value) => setFillColor({ ...fillColor(), b: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Alpha"}
-                    minMaxStep={[0,255,1]}
-                    value={fillColor().a}
-                    onChange={(value) => setFillColor({ ...fillColor(), a: value })}
-                />
-            </div>
-            )
-            //strokeColor
-            case 1: return (
-            <div>
-                <ui.NumberSliderUIElement 
-                    name={"Red"}
-                    minMaxStep={[0,255,1]}
-                    value={strokeColor().r}
-                    onChange={(value) => setStrokeColor({ ...strokeColor(), r: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Green"}
-                    minMaxStep={[0,255,1]}
-                    value={strokeColor().g}
-                    onChange={(value) => setStrokeColor({ ...strokeColor(), g: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Blue"}
-                    minMaxStep={[0,255,1]}
-                    value={strokeColor().b}
-                    onChange={(value) => setStrokeColor({ ...strokeColor(), b: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Alpha"}
-                    minMaxStep={[0,255,1]}
-                    value={strokeColor().a}
-                    onChange={(value) => setStrokeColor({ ...strokeColor(), a: value })}
-                />
-            </div>
-            )
-            //accentColor
-            case 2: return (
-            <div>
-                <ui.NumberSliderUIElement 
-                    name={"Red"}
-                    minMaxStep={[0,255,1]}
-                    value={accentColor().r}
-                    onChange={(value) => setAccentColor({ ...accentColor(), r: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Green"}
-                    minMaxStep={[0,255,1]}
-                    value={accentColor().g}
-                    onChange={(value) => setAccentColor({ ...accentColor(), g: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Blue"}
-                    minMaxStep={[0,255,1]}
-                    value={accentColor().b}
-                    onChange={(value) => setAccentColor({ ...accentColor(), b: value })}
-                />
-                <ui.NumberSliderUIElement 
-                    name={"Alpha"}
-                    minMaxStep={[0,255,1]}
-                    value={accentColor().a}
-                    onChange={(value) => setAccentColor({ ...accentColor(), a: value })}
-                />
-            </div>
-            )
         }
     }
     //Specific UI's from a toy
     function RenderSpecificUISettings() {
         if(toyType() == 0) return(<></>)
         else {
-            NewToy();
+            // NewToy();
             switch(toyType()) {
                 case 1: return (<GraviBoardUI channel={channel}></GraviBoardUI>);
                 case 2: return (<PolyDrumUI channel={channel}></PolyDrumUI>);
@@ -334,7 +155,6 @@ export default function SetupContainer( props: {channel: number}) {
             }
         }
     }
-
     function RenderToySelection() {
         return(
             <div class="flexList">
@@ -343,13 +163,6 @@ export default function SetupContainer( props: {channel: number}) {
                 <button id="thinButton" onClick={() => SetToyType(2)}>Poly Drum</button>
                 {/* <button id="thinButton" onClick={() => SetToyType(3)}>Square Keyboard</button> */}
             </div>
-        )
-    }
-    //Renders the Preset Managemtn UI system
-    function RenderPresetUI() {
-        NewToy();
-        return(
-            <PresetUI channel={channel}></PresetUI>
         )
     }
